listen method for everyone
